Added dice combination check and scoring system
Added a method called "checkDice" which returns an ArrayList of ArrayLists that can be described as a list of all possible combinations of the currently selectable dice.

Also added a method called "scoreCombination" which reads in a valid combination from the user and returns a score.

Both methods are a work in progress.